add abstract test classes for recoverable writers(s3/oss)
